# contract/openapi.yaml
openapi: 3.0.3
info:
  title: YourTJ v2 API
  version: 2.0.0
  description: >
    The single source of truth for the v2 HTTP contract. Client type bindings
    (TS / Swift / Dart) are generated from this file — change the API here first,
    then implement and regenerate. Conventions: bearer JWT auth; money/escrow
    writes use one-time signing intents and require `X-Wallet-Intent`,
    `X-Wallet-Sig`, and the same `Idempotency-Key`; cursor pagination on public
    lists; the error envelope `{ error: { code, message } }`; timestamps are Unix seconds.
servers:
  - url: https://api.yourtj.de/api/v2
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string }, description: Opaque pagination cursor }
    Limit: { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
...
  responses:
...
    Forbidden:    { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
    Page:
      type: object
      description: Base cursor-pagination envelope; typed lists override `items`.
      properties:
        items: { type: array, items: { type: object } }
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }

    # ---- identity ----
...
    AdminReasonInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
...
    AdminOverview:
      type: object
      required: [totalUsers, activeUsers, suspendedUsers, pendingReviewReports, pendingForumFlags, pendingDmReports, pendingMediaUploads, threadsToday, commentsToday, likesToday]
      properties:
        totalUsers: { type: integer, minimum: 0 }
        activeUsers: { type: integer, minimum: 0 }
        suspendedUsers: { type: integer, minimum: 0 }
        moderators: { type: integer, minimum: 0 }
        administrators: { type: integer, minimum: 0 }
        pendingReviewReports: { type: integer, minimum: 0 }
        pendingForumFlags: { type: integer, minimum: 0 }
        pendingDmReports: { type: integer, minimum: 0 }
        pendingMediaUploads: { type: integer, minimum: 0 }
        threadsToday: { type: integer, minimum: 0 }
        commentsToday: { type: integer, minimum: 0 }
        likesToday: { type: integer, minimum: 0 }
    AdminAuditEvent:
      type: object
      required: [id, actorKind, action, targetType, targetId, createdAt]
      properties:
        id: { type: string }
        actorKind: { type: string, enum: [account, system, service] }
        actorId: { type: string, nullable: true }
        actorHandle: { type: string, nullable: true }
        actorRole: { type: string, nullable: true }
        action: { type: string }
        targetType: { type: string }
        targetId: { type: string }
        reason: { type: string, nullable: true }
        metadata: { type: object, nullable: true }
        createdAt: { type: integer }
...
    AdminAuditEventPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminAuditEvent" } } } }
...
paths:
...
  /admin/overview:
    get:
      tags: [admin]
      summary: Administration dashboard counters
      responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/AdminOverview" } } } }, "403": { $ref: "#/components/responses/Forbidden" } }
  /admin/audit-events:
    get:
      tags: [admin]
      summary: Cursor-paginated immutable administration audit events
      parameters:
        - { name: actorId, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: targetType, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAuditEventPage" } } } }, "403": { $ref: "#/components/responses/Forbidden" } }
...
  /admin/selection/sync:
    post: { tags: [admin], summary: Trigger 一系统 sync job, requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AdminReasonInput" } } } }, responses: { "202": { description: queued } } }
  /admin/reviews/reindex:
    post: { tags: [admin], summary: Rebuild the Meilisearch review index, requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AdminReasonInput" } } } }, responses: { "202": { description: queued } } }
  /admin/forum/reindex:
    post: { tags: [admin], summary: Rebuild the Meilisearch forum index, requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AdminReasonInput" } } } }, responses: { "202": { description: queued } } }
...